massive refactor for proper packages
