+++ JwTikTakFragment/build.gradle
@@ -64,6 +64,4 @@ dependencies {
     implementation "com.jwplayer:jwplayer-ima:$JWPlayerVersion"
 
     implementation 'com.google.code.gson:gson:2.8.9'
-
-    implementation 'com.github.Jmilham21:ScrollingFeed:Tag'
 }
