+++ JwTikTakFragment/src/main/java/com/jmilham21/scrollingfeed/data/JwAdvertisement.kt
@@ -1,4 +1,4 @@
-package com.jwplayer.jwtiktak.data
+package com.jmilham21.scrollingfeed.data
 
 data class JwAdvertisement(
     var adUrl: String = ""
