+++ JwTikTakFragment/src/main/java/com/jmilham21/scrollingfeed/data/JwMedia.kt
@@ -1,0 +1,5 @@
+package com.jmilham21.scrollingfeed.data
+
+open class JwMedia() {
+
+}
