+++ JwTikTakFragment/src/main/java/com/jmilham21/scrollingfeed/fragment/VideoViewModel.kt
@@ -1,8 +1,8 @@
-package com.jwplayer.jwtiktak.fragment
+package com.jmilham21.scrollingfeed.fragment
 
 import androidx.lifecycle.ViewModel
-import com.jwplayer.jwtiktak.data.JwAdvertisement
-import com.jwplayer.jwtiktak.data.JwVideo
+import com.jmilham21.scrollingfeed.data.JwAdvertisement
+import com.jmilham21.scrollingfeed.data.JwVideo
 import com.jwplayer.pub.api.configuration.PlayerConfig
 import com.jwplayer.pub.api.configuration.UiConfig
 import com.jwplayer.pub.api.configuration.ads.ima.ImaAdvertisingConfig
