--- JwTikTakFragment/src/main/java/com/jwplayer/jwtiktak/data/JwMedia.kt
@@ -1,5 +1,0 @@
-package com.jwplayer.jwtiktak.data
-
-open class JwMedia() {
-
-}
