+++ settings.gradle
@@ -3,7 +3,6 @@ dependencyResolutionManagement {
     repositories {
         google()
         mavenCentral()
-        jcenter() // Warning: this repository is going to shut down soon
         maven {
             url 'https://mvn.jwplayer.com/content/repositories/releases/'
         }
